Добавляет код для модуля map.js, исправляет обработчик событий для движения главной иконки по карте
